Playground/backend/api/requirements.txt
# AONP API Server Dependencies

# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.1

# Models and validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment and configuration
python-dotenv==1.0.0

# CORS and security
slowapi==0.1.9  # Rate limiting

# Existing AONP dependencies
langchain==0.3.26
langchain-fireworks==0.2.10
langgraph==1.0.1
voyageai==0.5.0

# RAG System
chromadb==0.4.22
pymupdf==1.23.26

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0  # For async testing

# Monitoring (optional)
prometheus-client==0.19.0

